replace argparse with click and update example date
