+++ nested_sequence/text_classification/network_conf.py
@@ -6,12 +6,16 @@ def cnn_cov_group(group_input, hidden_size):
         input=group_input, context_len=3, hidden_size=hidden_size)
     conv4 = paddle.networks.sequence_conv_pool(
         input=group_input, context_len=4, hidden_size=hidden_size)
+
+    #output_group = paddle.layer.concat(input=[conv3, conv4])
+
     output_group = paddle.layer.fc(
         input=[conv3, conv4],
         size=hidden_size,
         param_attr=paddle.attr.ParamAttr(name='_cov_value_weight'),
         bias_attr=paddle.attr.ParamAttr(name='_cov_value_bias'),
         act=paddle.activation.Linear())
+
     return output_group
