+++ nested_sequence/text_classification/utils.py
@@ -1,61 +1,37 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-import logging
 import os
-import argparse
+import logging
 from collections import defaultdict
 
 logger = logging.getLogger("paddle")
 logger.setLevel(logging.INFO)
 
 
-def parse_train_cmd():
-    parser = argparse.ArgumentParser(
-        description="PaddlePaddle text classification demo")
-    parser.add_argument(
-        "--train_data_dir",
-        type=str,
-        required=False,
-        help=("path of training dataset (default: None). "
-              "if this parameter is not set, "
-              "imdb dataset will be used."),
-        default=None)
-    parser.add_argument(
-        "--test_data_dir",
-        type=str,
-        required=False,
-        help=("path of testing dataset (default: None). "
-              "if this parameter is not set, "
-              "imdb dataset will be used."),
-        default=None)
-    parser.add_argument(
-        "--word_dict",
-        type=str,
-        required=False,
-        help=("path of word dictionary (default: None)."
-              "if this parameter is not set, imdb dataset will be used."
-              "if this parameter is set, but the file does not exist, "
-              "word dictionay will be built from "
-              "the training data automatically."),
-        default=None)
-    parser.add_argument(
-        "--class_num",
-        type=int,
-        required=False,
-        help=("class number."),
-        default=2)
-    parser.add_argument(
-        "--batch_size",
-        type=int,
-        default=32,
-        help="the number of training examples in one forward/backward pass")
-    parser.add_argument(
-        "--num_passes", type=int, default=10, help="number of passes to train")
-    parser.add_argument(
-        "--model_save_dir",
-        type=str,
-        required=False,
-        help=("path to save the trained models."),
-        default="models")
+def build_dict(data_dir, save_path, use_col=1, cutoff_fre=1):
+    values = defaultdict(int)
 
-    return parser.parse_args()
+    for file_name in os.listdir(data_dir):
+        file_path = os.path.join(data_dir, file_name)
+        if not os.path.isfile(file_path):
+            continue
+        with open(file_path, "r") as fdata:
+            for line in fdata:
+                line_splits = line.strip().split("\t")
+                if len(line_splits) < use_col:
+                    continue
+                doc = line_splits[use_col]
+                for sent in doc.strip().split("."):
+                    for w in sent.split():
+                        values[w] += 1
+
+    values['<unk>'] = cutoff_fre
+    with open(save_path, "w") as f:
+        for v, count in sorted(
+                values.iteritems(), key=lambda x: x[1], reverse=True):
+            if count < cutoff_fre:
+                break
+            f.write("%s\t%d\n" % (v, count))
+
+
+def load_dict(dict_path):
+    return dict((line.strip().split("\t")[0], idx)
+                for idx, line in enumerate(open(dict_path, "r").readlines()))
